requirements2.txt
a2wsgi==1.10.8
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-google==10.1.0
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-standard==1.0.0
apache-airflow-task-sdk==1.0.0
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
cadwyn==5.3.3
cattrs==24.1.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==41.0.7
datasets==3.5.0
db-dtypes==1.4.2
deprecated==1.2.18
dill==0.3.8
dnspython==2.7.0
docstring-parser==0.16
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flask==3.1.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2024.12.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gitdb==4.0.12
gitpython==3.1.44
google-ads==26.1.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.90.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-automl==2.16.3
google-cloud-bigquery==3.31.0
google-cloud-bigquery-datatransfer==3.19.1
google-cloud-bigtable==2.30.1
google-cloud-build==3.31.1
google-cloud-compute==1.30.0
google-cloud-container==2.56.1
google-cloud-core==2.4.3
google-cloud-datacatalog==3.26.1
google-cloud-dataflow-client==0.8.17
google-cloud-dataform==0.6.1
google-cloud-dataplex==2.10.1
google-cloud-dataproc==5.18.1
google-cloud-dataproc-metastore==1.18.2
google-cloud-dlp==3.29.0
google-cloud-kms==3.4.1
google-cloud-language==2.17.1
google-cloud-logging==3.12.1
google-cloud-memcache==1.12.1
google-cloud-monitoring==2.27.1
google-cloud-orchestration-airflow==1.17.5
google-cloud-os-login==2.17.1
google-cloud-pubsub==2.29.0
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-spanner==3.53.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-tasks==2.19.2
google-cloud-texttospeech==2.26.0
google-cloud-translate==3.20.2
google-cloud-videointelligence==2.16.1
google-cloud-vision==3.10.1
google-cloud-workflows==1.18.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.71.0
grpcio-gcp==0.2.2
grpcio-status==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
json-merge-patch==0.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.11.0
libcst==1.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.4.0
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.21.1
onnxruntime-tools==1.7.0
opentelemetry-api==1.32.1
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-flask==0.53b1
opentelemetry-instrumentation-requests==0.53b1
opentelemetry-instrumentation-wsgi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
optimum==1.24.0
packaging==25.0
pandas==2.2.3
pandas-gbq==0.28.0
pathspec==0.12.1
pendulum==3.1.0
pillow==11.2.1
pluggy==1.5.0
portalocker==2.10.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydata-google-auth==1.9.1
pygments==2.19.1
pyjwt==2.10.1
pymongo==4.12.0
pyopenssl==25.0.0
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.13.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retryhttp==1.3.2
rich==14.0.0
rich-argparse==1.7.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setproctitle==1.3.5
setuptools==66.1.1
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==1.4.54
sqlalchemy-bigquery==1.13.0
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.2.0
svcs==25.1.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
thrift==0.21.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uc-micro-py==1.0.3
universal-pathlib==0.2.6
uritemplate==4.1.1
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
